add code to allow for 4.x upgrade debugging
git-svn-id: file:///Users/michaelkoepke/svn/semiologic/themes/semiologic/trunk@646 6c9c31c4-cb3d-4424-af3d-876279169c7e

+++ inc/upgrade/options.php
@@ -60,8 +60,11 @@ function sem_clean_up_options()
 	delete_option('sm_cpages');							#arne sitemap
 	
 	delete_option('sem_widget_contexts');
-	delete_option('sem5_nav');
-	delete_option('semiologic');
+	if ( !defined('sem_install_test') )
+	{
+		delete_option('sem5_nav');
+		delete_option('semiologic');
+	}
 	
 	# clean up cache
 	foreach ( glob(ABSPATH . "wp-content/cache/yt-*") as $cache_file )
